updating db search templates

+++ trunk/OGen-Kick/src/Kick-businesslayer/mod-credentials/SBO_CRD_Profile.cs
@@ -725,11 +725,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_CRD_Profile[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
@@ -789,11 +789,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_CRD_ProfileProfile[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
@@ -970,11 +970,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vCRD_UserProfile[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
@@ -1034,11 +1034,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vCRD_ProfilePermition[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
