+++ trunk/OGen-Kick/src/Kick-businesslayer/mod-dic/SBO_DIC_Dic.cs
@@ -699,11 +699,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vDIC_ApplicationLanguage[] _output = null;
 
 			List<int> _errorlist;
@@ -749,11 +749,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vDIC_Language[] _output = null;
 
 			List<int> _errorlist;
@@ -800,11 +800,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vDIC_Language[] _output = null;
 
 			List<int> _errorlist;
