+++ trunk/OGen-Kick/src/Kick-businesslayer/mod-log/SBO_LOG_Log.cs
@@ -374,11 +374,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out,
+			out long page_itemsCount_out,
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			List<int> _errorlist;
 			Guid _sessionguid;
 			Sessionuser _sessionuser;
