+++ trunk/OGen-Kick/src/Kick-businesslayer/mod-news/SBO_NWS_Profile.cs
@@ -719,11 +719,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vNWS_Profile[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
@@ -781,11 +781,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_vNWS_Profile[] _output = null;
 			List<int> _errorlist;
 
@@ -848,11 +848,11 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in,
 			int page_in,
 			int page_numRecords_in,
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 
 			out int[] errors_out
 		) {
-			page_itemsCount_out = -1;
+			page_itemsCount_out = -1L;
 			SO_NWS_ContentProfile[] _output = null;
 			List<int> _errorlist;
 			Guid _sessionguid;
