+++ trunk/OGen-Kick/src/Kick-web/admin/CRD-Profile-list.aspx.cs
@@ -55,7 +55,7 @@ namespace OGen.NTier.Kick.presentationlayer.weblayer {
 		#region public void Bind();
 		public void Bind() {
 			int[] _errors;
-			int _count;
+			long _count;
 			SO_CRD_Profile[] _authors
 				= BusinessInstances.CRD_Profile.InstanceClient.getRecord_all(
 					utils.User.SessionGuid,
